apiserver/requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
aiomysql==0.0.20
python-dateutil==2.8.1
PyPika==0.37.6
tortoise-orm==0.16.10
gunicorn==20.0.4
python-json-logger==0.1.11
deserialize==1.8.0
motor==2.0.0
jasyncq==0.1.2